comments out GPS tests (for now); added /api/pilot/getAll endpoint
i can't get the gps tests working at the moment. the gps service
relies on the current time and date to check whether or not the ship
is in the permitted range of entry. the tests are unable to place
orders of the current time with the guarantee that they'll be
confirmed due to the checks against the current tide and the pilot's
schedule.